+++ src/hooks/api/use-stripe-connect.ts
@@ -3,11 +3,15 @@
  * Calls the stripe-connect Supabase Edge Function directly (no NestJS).
  */
 import { useMutation, useQuery, useQueryClient, mutationOptions } from '@tanstack/react-query'
-import { QUERY_CACHE_TIMES } from '#lib/constants/query-config'
 import { handleMutationError } from '#lib/mutation-error-handler'
-import { createClient } from '#lib/supabase/client'
 import { mutationKeys } from './mutation-keys'
-import type { ConnectedAccountWithIdentity } from '#types/stripe'
+import {
+	stripeConnectKeys,
+	stripeConnectQueries,
+	callStripeConnectFunction,
+} from './query-keys/stripe-connect-keys'
+
+export { stripeConnectKeys, stripePayoutKeys, type Payout, type Transfer } from './query-keys/stripe-connect-keys'
 
 // ============================================================================
 // Types
@@ -18,99 +22,6 @@ interface CreateConnectAccountRequest {
 	businessName?: string
 	country?: string
 	entityType?: 'individual' | 'company'
-}
-
-interface BalanceAmount {
-	amount: number
-	currency: string
-}
-
-interface BalanceResponse {
-	balance: {
-		available: BalanceAmount[]
-		pending: BalanceAmount[]
-	}
-}
-
-export interface Payout {
-	id: string
-	amount: number
-	currency: string
-	status: string
-	arrival_date: number
-	created: number
-	method: string
-	type: string
-	description?: string
-	failure_message?: string
-}
-
-interface PayoutsResponse {
-	payouts: Payout[]
-	hasMore: boolean
-}
-
-export interface Transfer {
-	id: string
-	amount: number
-	currency: string
-	created: number
-	description?: string
-	metadata?: Record<string, string>
-}
-
-interface TransfersResponse {
-	transfers: Transfer[]
-	hasMore: boolean
-}
-
-// ============================================================================
-// Query Keys
-// ============================================================================
-
-export const stripeConnectKeys = {
-	all: ['stripeConnect'] as const,
-	account: () => [...stripeConnectKeys.all, 'account'] as const,
-}
-
-export const stripePayoutKeys = {
-	all: ['stripePayouts'] as const,
-	balance: () => [...stripePayoutKeys.all, 'balance'] as const,
-	payouts: (params?: { limit?: number; starting_after?: string }) =>
-		[...stripePayoutKeys.all, 'list', params] as const,
-	transfers: (params?: { limit?: number; starting_after?: string }) =>
-		[...stripePayoutKeys.all, 'transfers', params] as const,
-}
-
-// ============================================================================
-// Edge Function helper
-// ============================================================================
-
-async function callStripeConnectFunction<T>(
-	action: string,
-	body?: Record<string, unknown>
-): Promise<T> {
-	const supabase = createClient()
-	const { data: sessionData } = await supabase.auth.getSession()
-	const token = sessionData.session?.access_token
-	if (!token) throw new Error('Not authenticated')
-
-	const baseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL
-	const response = await fetch(`${baseUrl}/functions/v1/stripe-connect`, {
-		method: 'POST',
-		headers: {
-			Authorization: `Bearer ${token}`,
-			'Content-Type': 'application/json',
-		},
-		body: JSON.stringify({ action, ...body }),
-	})
-
-	if (!response.ok) {
-		const err = await response.json().catch(() => ({ error: response.statusText })) as Record<string, unknown>
-		throw new Error((err.error as string | undefined) ?? `stripe-connect failed: ${response.status}`)
-	}
-
-	return response.json() as Promise<T>
 }
 
 // ============================================================================
@@ -160,22 +71,7 @@ const stripeConnectMutationFactories = {
 // ============================================================================
 
 export function useConnectedAccount() {
-	return useQuery({
-		queryKey: stripeConnectKeys.account(),
-		queryFn: async (): Promise<ConnectedAccountWithIdentity | null> => {
-			try {
-				const result = await callStripeConnectFunction<{
-					account: ConnectedAccountWithIdentity | null
-					hasAccount: boolean
-				}>('account')
-				return result.account
-			} catch {
-				return null
-			}
-		},
-		...QUERY_CACHE_TIMES.DETAIL,
-		retryOnMount: false,
-	})
+	return useQuery(stripeConnectQueries.account())
 }
 
 export function useCreateConnectedAccountMutation() {
@@ -226,40 +122,19 @@ export function useStripeDashboardLink() {
 }
 
 export function useConnectedAccountBalance() {
-	return useQuery({
-		queryKey: stripePayoutKeys.balance(),
-		queryFn: async (): Promise<BalanceResponse['balance']> => {
-			const result = await callStripeConnectFunction<BalanceResponse>('balance')
-			return result.balance
-		},
-		...QUERY_CACHE_TIMES.STATS,
-	})
+	return useQuery(stripeConnectQueries.balance())
 }
 
 export function useConnectedAccountPayouts(params?: {
 	limit?: number
 	starting_after?: string
 }) {
-	return useQuery({
-		queryKey: stripePayoutKeys.payouts(params),
-		queryFn: async (): Promise<{ payouts: Payout[]; hasMore: boolean }> => {
-			const result = await callStripeConnectFunction<PayoutsResponse>('payouts', params)
-			return { payouts: result.payouts, hasMore: result.hasMore }
-		},
-		...QUERY_CACHE_TIMES.LIST,
-	})
+	return useQuery(stripeConnectQueries.payouts(params))
 }
 
 export function useConnectedAccountTransfers(params?: {
 	limit?: number
 	starting_after?: string
 }) {
-	return useQuery({
-		queryKey: stripePayoutKeys.transfers(params),
-		queryFn: async (): Promise<{ transfers: Transfer[]; hasMore: boolean }> => {
-			const result = await callStripeConnectFunction<TransfersResponse>('transfers', params)
-			return { transfers: result.transfers, hasMore: result.hasMore }
-		},
-		...QUERY_CACHE_TIMES.LIST,
-	})
+	return useQuery(stripeConnectQueries.transfers(params))
 }
